Relocate states.py
